Move file storage tasks away from api.go
[noissue]
